Implement (de)serialization functions for Inventory-QuiverComponent.

+++ src/systems/DragAndDropSystem.cpp
@@ -23,8 +23,8 @@ void DragAndDropSystem::update(entityx::EntityManager&, entityx::EventManager&, 
 		//Compute the drag and drop data
 		float delta_x{m_line[1].position.x- m_line[0].position.x};
 		float delta_y{m_line[1].position.y- m_line[0].position.y};
-		float power{hypot(delta_x, delta_y)};//Distance between the two points
-		float angle{atan2(delta_x, delta_y)};//Angle of the line with the horizontal axis
+		double power{hypot(delta_x, delta_y)};//Distance between the two points
+		double angle{atan2(delta_x, delta_y)};//Angle of the line with the horizontal axis
 		angle += b2_pi/2.f;//Turn the angle of 90 degrees to fit with the gameplay requirements
 //		if(angle > b2_pi + b2_pi/4.f)//Keep the angle in the range [-pi, pi]
 //			angle = angle - 2*b2_pi;
@@ -46,7 +46,7 @@ void DragAndDropSystem::setDragAndDropActivation(bool isActive)
 	{
 		float delta_x{m_line[1].position.x- m_line[0].position.x};
 		float delta_y{m_line[1].position.y- m_line[0].position.y};
-		float angle{atan2(delta_x, delta_y) + b2_pi/2.f};//Angle of the line with the horizontal axis
+		double angle{atan2(delta_x, delta_y) + b2_pi/2.f};//Angle of the line with the horizontal axis
 		Command bendCommand;
 		bendCommand.targetIsSpecific = false;
 		bendCommand.category = Category::Player;
